cleaning up before branch push

+++ src/colormatrix.cpp
@@ -140,7 +140,7 @@ void KmerMatrix::get_row(uint64_t row_idx, vector<uint64_t>& to_fill)
 
 __uint128_t convert_uint64_to_uint128(uint64_t r) {
 	__uint128_t out = 0;
-	out |= r;
+	out |= (__uint128_t)r;
 	out<<64;
     //return ((__uint128_t)0 << 0) | ((__uint128_t)r << 64);//works in mini, but not ec
 	return out;
